+++ test/bigtest/interactors/settings-display-interactor.js
@@ -1,3 +1,0 @@
-import { Interactor } from '@bigtest/interactor';
-
-export const settingsDisplayInteractor = new Interactor('#paneHeader2-pane-title');