+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## 1.0.0 (IN PROGRESS)
 * UI app created. Refs UIDEXP-8.
 * Handle security issue because of eslint version, remove excess eslint rules, upgrade babel and eslint-config-stripes versions. Refs UIDEXP-17.
-* Add dependency on @folio/stripes-data-transfer-components. Refs UIDEXP-11.
+* Add dependency on @folio/stripes-data-transfer-components. Add file uploader feature on landing page. Refs UIDEXP-11.